+++ command/commands.go
@@ -76,6 +76,8 @@ const (
 	flagNameListingVisibility = "listing-visibility"
 	// flagNamePassthroughRequestHeaders is the flag name used to set passthrough request headers to the backend
 	flagNamePassthroughRequestHeaders = "passthrough-request-headers"
+	// flagNameAllowedResponseHeaders is used to set allowed response headers from a plugin
+	flagNameAllowedResponseHeaders = "allowed-response-headers"
 	// flagNameTokenType is the flag name used to force a specific token type
 	flagNameTokenType = "token-type"
 )
@@ -420,6 +422,16 @@ func initCommands(ui, serverCmdUi cli.Ui, runOpts *RunOptions) {
 		},
 		"policy write": func() (cli.Command, error) {
 			return &PolicyWriteCommand{
+				BaseCommand: getBaseCommand(),
+			}, nil
+		},
+		"print": func() (cli.Command, error) {
+			return &PrintCommand{
+				BaseCommand: getBaseCommand(),
+			}, nil
+		},
+		"print token": func() (cli.Command, error) {
+			return &PrintTokenCommand{
 				BaseCommand: getBaseCommand(),
 			}, nil
 		},