+++ helper/testhelpers/testhelpers.go
@@ -5,13 +5,13 @@ import (
 	"errors"
 	"fmt"
 	"math/rand"
+	"net/http"
 	"sync"
 	"time"
 
 	log "github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/vault/helper/consts"
 	"github.com/hashicorp/vault/helper/xor"
-	vaulthttp "github.com/hashicorp/vault/http"
 	"github.com/hashicorp/vault/physical"
 	"github.com/hashicorp/vault/physical/inmem"
 	"github.com/hashicorp/vault/vault"
@@ -28,8 +28,12 @@ type ReplicatedTestClusters struct {
 func (r *ReplicatedTestClusters) Cleanup() {
 	r.PerfPrimaryCluster.Cleanup()
 	r.PerfSecondaryCluster.Cleanup()
-	r.PerfPrimaryDRCluster.Cleanup()
-	r.PerfSecondaryDRCluster.Cleanup()
+	if r.PerfPrimaryDRCluster != nil {
+		r.PerfPrimaryDRCluster.Cleanup()
+	}
+	if r.PerfSecondaryDRCluster != nil {
+		r.PerfSecondaryDRCluster.Cleanup()
+	}
 }
 
 // Generates a root token on the target cluster.
@@ -100,34 +104,77 @@ func RandomWithPrefix(name string) string {
 	return fmt.Sprintf("%s-%d", name, rand.New(rand.NewSource(time.Now().UnixNano())).Int())
 }
 
+func EnsureCoresSealed(t testing.T, c *vault.TestCluster) {
+	t.Helper()
+	for _, core := range c.Cores {
+		EnsureCoreSealed(t, core)
+	}
+}
+
+func EnsureCoreSealed(t testing.T, core *vault.TestClusterCore) error {
+	core.Seal(t)
+	timeout := time.Now().Add(60 * time.Second)
+	for {
+		if time.Now().After(timeout) {
+			return fmt.Errorf("timeout waiting for core to seal")
+		}
+		if core.Core.Sealed() {
+			break
+		}
+		time.Sleep(250 * time.Millisecond)
+	}
+	return nil
+}
+
 func EnsureCoresUnsealed(t testing.T, c *vault.TestCluster) {
 	t.Helper()
 	for _, core := range c.Cores {
-		if !core.Sealed() {
-			continue
-		}
+		EnsureCoreUnsealed(t, c, core)
+	}
+}
+func EnsureCoreUnsealed(t testing.T, c *vault.TestCluster, core *vault.TestClusterCore) {
+	if !core.Sealed() {
+		return
+	}
 
-		client := core.Client
-		client.Sys().ResetUnsealProcess()
-		for j := 0; j < len(c.BarrierKeys); j++ {
-			statusResp, err := client.Sys().Unseal(base64.StdEncoding.EncodeToString(c.BarrierKeys[j]))
-			if err != nil {
-				// Sometimes when we get here it's already unsealed on its own
-				// and then this fails for DR secondaries so check again
-				if core.Sealed() {
-					t.Fatal(err)
-				}
-				break
+	client := core.Client
+	client.Sys().ResetUnsealProcess()
+	for j := 0; j < len(c.BarrierKeys); j++ {
+		statusResp, err := client.Sys().Unseal(base64.StdEncoding.EncodeToString(c.BarrierKeys[j]))
+		if err != nil {
+			// Sometimes when we get here it's already unsealed on its own
+			// and then this fails for DR secondaries so check again
+			if core.Sealed() {
+				t.Fatal(err)
 			}
-			if statusResp == nil {
-				t.Fatal("nil status response during unseal")
-			}
-			if !statusResp.Sealed {
-				break
-			}
+			break
 		}
-		if core.Sealed() {
-			t.Fatal("core is still sealed")
+		if statusResp == nil {
+			t.Fatal("nil status response during unseal")
+		}
+		if !statusResp.Sealed {
+			break
+		}
+	}
+	if core.Sealed() {
+		t.Fatal("core is still sealed")
+	}
+}
+
+func EnsureCoreIsPerfStandby(t testing.T, core *vault.TestClusterCore) {
+	t.Helper()
+	start := time.Now()
+	for {
+		health, err := core.Client.Sys().Health()
+		if err != nil {
+			t.Fatal(err)
+		}
+		if health.PerformanceStandby {
+			break
+		}
+		time.Sleep(time.Millisecond * 500)
+		if time.Now().After(start.Add(time.Second * 60)) {
+			t.Fatal("did not become a perf standby")
 		}
 	}
 }
@@ -146,7 +193,7 @@ func WaitForReplicationState(t testing.T, c *vault.Core, state consts.Replicatio
 	}
 }
 
-func GetClusterAndCore(t testing.T, logger log.Logger) (*vault.TestCluster, *vault.TestClusterCore) {
+func GetClusterAndCore(t testing.T, logger log.Logger, handlerFunc func(*vault.HandlerProperties) http.Handler) (*vault.TestCluster, *vault.TestClusterCore) {
 	inm, err := inmem.NewTransactionalInmem(nil, logger)
 	if err != nil {
 		t.Fatal(err)
@@ -162,7 +209,7 @@ func GetClusterAndCore(t testing.T, logger log.Logger) (*vault.TestCluster, *vau
 	}
 
 	cluster := vault.NewTestCluster(t, coreConfig, &vault.TestClusterOptions{
-		HandlerFunc: vaulthttp.Handler,
+		HandlerFunc: handlerFunc,
 		Logger:      logger,
 	})
 	cluster.Start()
@@ -175,7 +222,7 @@ func GetClusterAndCore(t testing.T, logger log.Logger) (*vault.TestCluster, *vau
 	return cluster, core
 }
 
-func GetFourReplicatedClusters(t testing.T) *ReplicatedTestClusters {
+func GetPerfReplicatedClusters(t testing.T, handlerFunc func(*vault.HandlerProperties) http.Handler) *ReplicatedTestClusters {
 	ret := &ReplicatedTestClusters{}
 
 	logger := log.New(&log.LoggerOptions{
@@ -185,13 +232,37 @@ func GetFourReplicatedClusters(t testing.T) *ReplicatedTestClusters {
 	// Set this lower so that state populates quickly to standby nodes
 	vault.HeartbeatInterval = 2 * time.Second
 
-	ret.PerfPrimaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-pri"))
+	ret.PerfPrimaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-pri"), handlerFunc)
 
-	ret.PerfSecondaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-sec"))
+	ret.PerfSecondaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-sec"), handlerFunc)
 
-	ret.PerfPrimaryDRCluster, _ = GetClusterAndCore(t, logger.Named("perf-pri-dr"))
+	SetupTwoClusterPerfReplication(t, ret.PerfPrimaryCluster, ret.PerfSecondaryCluster)
 
-	ret.PerfSecondaryDRCluster, _ = GetClusterAndCore(t, logger.Named("perf-sec-dr"))
+	// Wait until poison pills have been read
+	time.Sleep(45 * time.Second)
+	EnsureCoresUnsealed(t, ret.PerfPrimaryCluster)
+	EnsureCoresUnsealed(t, ret.PerfSecondaryCluster)
+
+	return ret
+}
+
+func GetFourReplicatedClusters(t testing.T, handlerFunc func(*vault.HandlerProperties) http.Handler) *ReplicatedTestClusters {
+	ret := &ReplicatedTestClusters{}
+
+	logger := log.New(&log.LoggerOptions{
+		Mutex: &sync.Mutex{},
+		Level: log.Trace,
+	})
+	// Set this lower so that state populates quickly to standby nodes
+	vault.HeartbeatInterval = 2 * time.Second
+
+	ret.PerfPrimaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-pri"), handlerFunc)
+
+	ret.PerfSecondaryCluster, _ = GetClusterAndCore(t, logger.Named("perf-sec"), handlerFunc)
+
+	ret.PerfPrimaryDRCluster, _ = GetClusterAndCore(t, logger.Named("perf-pri-dr"), handlerFunc)
+
+	ret.PerfSecondaryDRCluster, _ = GetClusterAndCore(t, logger.Named("perf-sec-dr"), handlerFunc)
 
 	SetupFourClusterReplication(t, ret.PerfPrimaryCluster, ret.PerfSecondaryCluster, ret.PerfPrimaryDRCluster, ret.PerfSecondaryDRCluster)
 
@@ -203,6 +274,46 @@ func GetFourReplicatedClusters(t testing.T) *ReplicatedTestClusters {
 	EnsureCoresUnsealed(t, ret.PerfSecondaryDRCluster)
 
 	return ret
+}
+
+func SetupTwoClusterPerfReplication(t testing.T, perfPrimary, perfSecondary *vault.TestCluster) {
+	// Enable performance primary
+	_, err := perfPrimary.Cores[0].Client.Logical().Write("sys/replication/performance/primary/enable", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	WaitForReplicationState(t, perfPrimary.Cores[0].Core, consts.ReplicationPerformancePrimary)
+
+	// get performance token
+	secret, err := perfPrimary.Cores[0].Client.Logical().Write("sys/replication/performance/primary/secondary-token", map[string]interface{}{
+		"id": "1",
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	token := secret.WrapInfo.Token
+
+	// enable performace secondary
+	secret, err = perfSecondary.Cores[0].Client.Logical().Write("sys/replication/performance/secondary/enable", map[string]interface{}{
+		"token":   token,
+		"ca_file": perfPrimary.CACertPEMFile,
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	WaitForReplicationState(t, perfSecondary.Cores[0].Core, consts.ReplicationPerformanceSecondary)
+	time.Sleep(time.Second * 3)
+	perfSecondary.BarrierKeys = perfPrimary.BarrierKeys
+
+	EnsureCoresUnsealed(t, perfSecondary)
+	rootToken := GenerateRoot(t, perfSecondary, false)
+	perfSecondary.Cores[0].Client.SetToken(rootToken)
+	for _, core := range perfSecondary.Cores {
+		core.Client.SetToken(rootToken)
+	}
 }
 
 func SetupFourClusterReplication(t testing.T, perfPrimary, perfSecondary, perfDRSecondary, perfSecondaryDRSecondary *vault.TestCluster) {
@@ -269,7 +380,9 @@ func SetupFourClusterReplication(t testing.T, perfPrimary, perfSecondary, perfDR
 
 	EnsureCoresUnsealed(t, perfSecondary)
 	rootToken := GenerateRoot(t, perfSecondary, false)
-	perfSecondary.Cores[0].Client.SetToken(rootToken)
+	for _, core := range perfSecondary.Cores {
+		core.Client.SetToken(rootToken)
+	}
 
 	// Enable dr primary on perf secondary
 	_, err = perfSecondary.Cores[0].Client.Logical().Write("sys/replication/dr/primary/enable", nil)